Extracted serial interface handling from modbus-rtu*
