+++ src/modbus-rtu.h
@@ -19,6 +19,11 @@ MODBUS_BEGIN_DECLS
 MODBUS_API modbus_t* modbus_new_rtu(const char *device, int baud, char parity,
                                     int data_bit, int stop_bit);
 
+/*
+ * The get/set serial mode and get/set rts methods should be deprecated, 
+ * as they just forward to the serial backend.
+ * Currently they are kept so the known interface does not change.
+ */ 
 #define MODBUS_RTU_RS232 0
 #define MODBUS_RTU_RS485 1
